+++ coders/mvg.c
@@ -171,8 +171,8 @@ static Image *ReadMVGImage(const ImageInfo *image_info,ExceptionInfo *exception)
           continue;
         (void) sscanf(primitive,"viewbox %lf %lf %lf %lf",&bounds.x1,
           &bounds.y1,&bounds.x2,&bounds.y2);
-        image->columns= (bounds.x2-bounds.x1+0.5);
-        image->rows= (bounds.y2-bounds.y1+0.5);
+        image->columns=(unsigned long) (bounds.x2-bounds.x1+0.5);
+        image->rows=(unsigned long) (bounds.y2-bounds.y1+0.5);
         break;
       }
     }
